Split QuerySetup into QuerySetup & QueryOutput, refactor query out sel
- QuerySetup split into QuerySetup & QueryOutput (also in accordance to corresponding test files);
- move output (sink) selection from QueryCommand entirely into QueryOutput.
